requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
Pillow>=10.0
PyYAML>=6.0
opencv-python>=4.8
matplotlib>=3.8
tqdm>=4.66
torch>=2.2
torchvision>=0.17
